# Keyword/Gene Count Query Enhancement