This PR updates the query for counting cvterms associated to expression analyses. 

### Benchamarks
| Orgnaism | # of terms |  Previous Query | New Query |
| --------- | ----------- | ---------------- | ----------- |
| Green Ash | 5 | 1.1 mins | 5.55 secs |
| Northern Red Oak | 1 | 6 secs | 4.38 secs | 

As you can see above, small changes in the number of terms can make a huge difference in performance between the old and the new query.